refactor(landing): decompose the 865-line Landing.jsx into sections/ + shared modules
Landing.jsx had every section (header, hero, problem, ritual, film-file, briefing,
DNA, community, M.'s letter, pricing, final CTA, footer) plus three shared helpers
and the PICKS dataset in one 865-line file. Split it for manageability — and to set
the decomposition pattern for every feature surface:

  landing/
  ├── Landing.jsx     # slim composition shell (865 → 53 LOC)
  ├── data.js         # TMDB() + PICKS[] (shared film dataset)
  ├── primitives.jsx  # Reveal, Poster, Stars (shared landing helpers)
  └── sections/       # one file per section (12)

PURE extraction — verbatim code move, no behaviour or style change. Each section
imports C/GRAD from tokens, PICKS from ../data, and primitives from ../primitives;
landing.css stays a single global import in Landing.jsx. The router still
lazy-imports the unchanged default export.

Convention (now in CLAUDE.md): editorial / many-section surfaces use a sections/
subfolder; data-driven app surfaces keep sections-top/bottom. Both: the entry
composes, section bodies live in their own files.

Render is byte-identical: full-page screenshot of / (stabilized like the visual
test) is md5-identical before vs after. lint clean · test 417 · build ✓.
